refactoring partial tracked reads
